requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
langchain-ollama>=0.1.0
